requirements.txt

customtkinter
packaging
pillow
pyinstaller
pytest
pytest-mock
pytest-cov
bibtexparser>=1.4.0
